app/permalink/requirements.txt
Django==5.2
djangorestframework==3.16.0
Markdown==3.7
django-filter==25.1
